cmd/*,tui/*: Rename 'ui' to 'tui'
GitHub-Closes: #55
